# Speed up default_heap_size