`System::new_all` does a lot more than just measuring available mem, so let's cut those extra syscalls to make startup faster.

Bench (examples/hello):

Old:

```
$ hyperfine "target/release/main run '(Main 0)'"
Benchmark 1: target/release/main run '(Main 0)'
  Time (mean ± σ):     513.4 ms ±  36.6 ms    [User: 285.6 ms, System: 455.2 ms]
  Range (min … max):   469.5 ms … 587.2 ms    10 runs
```

New:

```
$ hyperfine "target/release/main run '(Main 0)'"
Benchmark 1: target/release/main run '(Main 0)'
  Time (mean ± σ):      77.5 ms ±   4.9 ms    [User: 18.0 ms, System: 59.9 ms]
  Range (min … max):    67.2 ms …  85.6 ms    33 runs
```